requirements.txt
# FastAPI + Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Templates
jinja2>=3.1.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Sessions
itsdangerous>=2.2.0

# Payments (optional)
stripe>=10.0.0

# Dev
python-dotenv>=1.0.0
